Dummy task test refactored and moved to separate package
