+++ tests/golem/task/dummy/computation.py
@@ -1,4 +1,4 @@
-# Script that computes a subtask of a 'dummy task'
+# Functions used in the computation of subtasks of the dummy task
 
 import hashlib
 import random
@@ -26,7 +26,7 @@ def find_pow(input_data, difficulty, result_size):
     :param int result_size:
     :rtype long:
     """
-    solution = random.getrandbits(result_size * 8)
+    solution = random.getrandbits(result_size * 4)
     while True:
         if check_pow(solution, input_data, difficulty):
             return solution
@@ -41,20 +41,18 @@ def run_dummy_task(data_file, subtask_data, difficulty, result_size):
     :param str subtask_data: subtask-specific part of data
     :param int difficulty: required difficulty
     :param int result_size: size of the solution string S
-    :rtype DummyTaskResult"""
-    # TODO: setup logging
-    print 'Running dummy task, data_file = ', data_file, 'result_size = ', result_size, 'difficulty = %x' % difficulty
+    :rtype DummyTaskResult\
+    """
+    print '[DUMMY TASK] computation started, data_file = ', data_file, \
+          ', result_size = ', result_size, \
+          ', difficulty = 0x%08x' % difficulty
     t0 = time.clock()
 
-    with open(data_file, 'rb') as f:
+    with open(data_file, 'r') as f:
         shared_input = f.read()
     solution = find_pow(shared_input + subtask_data, difficulty, result_size)
+    result = '%x' % solution
+    assert len(result) == result_size
 
-    print 'Task computation finished, time =', time.clock() - t0, ' sec'
-    return { 'data' : '%x' % solution, 'result_type' : 0 }
-
-
-
-
-
-
+    print '[DUMMY TASK] computation finished, time =', time.clock() - t0, 'sec'
+    return {'data': result, 'result_type': 0}
