+++ tests/golem/task/dummy/logging.ini
@@ -1,0 +1,38 @@
+[loggers]
+keys=root, peewee, pystun
+
+[handlers]
+keys=consoleHandler, fileHandler
+
+[formatters]
+keys=simpleFormatter
+
+[logger_root]
+level=DEBUG
+handlers=consoleHandler
+
+[logger_peewee]
+level=INFO
+handlers=consoleHandler, fileHandler
+qualname=peewee
+
+[logger_pystun]
+level=INFO
+handlers=consoleHandler, fileHandler
+qualname=pystun
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatter
+args=(sys.stdout,)
+
+[handler_fileHandler]
+class=FileHandler
+level=DEBUG
+formatter=simpleFormatter
+args=('golem.log',)
+
+[formatter_simpleFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
+datefmt=
